rename: ErrorCode, Custom exception 상위 패키지로 이동
