+++ api/src/main/java/com/programmers/bucketback/domains/item/application/ItemEnrollValidator.java
@@ -3,8 +3,8 @@ package com.programmers.bucketback.domains.item.application;
 import org.springframework.stereotype.Component;
 
 import com.programmers.bucketback.domains.item.repository.ItemRepository;
-import com.programmers.bucketback.error.exception.BusinessException;
-import com.programmers.bucketback.error.exception.ErrorCode;
+import com.programmers.bucketback.error.BusinessException;
+import com.programmers.bucketback.error.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
