+++ api/src/main/java/com/programmers/bucketback/global/error/ErrorResponse.java
@@ -7,7 +7,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.programmers.bucketback.error.exception.ErrorCode;
+import com.programmers.bucketback.error.ErrorCode;
 
 import lombok.AccessLevel;
 import lombok.Getter;
