+++ common/src/main/java/com/programmers/bucketback/error/BusinessException.java
@@ -1,4 +1,4 @@
-package com.programmers.bucketback.error.exception;
+package com.programmers.bucketback.error;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
