+++ domain/src/main/java/com/programmers/bucketback/domains/member/domain/vo/Introduction.java
@@ -1,7 +1,7 @@
 package com.programmers.bucketback.domains.member.domain.vo;
 
-import com.programmers.bucketback.error.exception.BusinessException;
-import com.programmers.bucketback.error.exception.ErrorCode;
+import com.programmers.bucketback.error.BusinessException;
+import com.programmers.bucketback.error.ErrorCode;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
