+++ domain/src/main/java/com/programmers/bucketback/domains/member/domain/vo/Nickname.java
@@ -3,8 +3,8 @@ package com.programmers.bucketback.domains.member.domain.vo;
 import java.util.Objects;
 import java.util.regex.Pattern;
 
-import com.programmers.bucketback.error.exception.BusinessException;
-import com.programmers.bucketback.error.exception.ErrorCode;
+import com.programmers.bucketback.error.BusinessException;
+import com.programmers.bucketback.error.ErrorCode;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
