+++ domain/src/main/java/com/programmers/bucketback/domains/vote/model/VoteStatusCondition.java
@@ -6,8 +6,8 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import com.programmers.bucketback.error.exception.BusinessException;
-import com.programmers.bucketback.error.exception.ErrorCode;
+import com.programmers.bucketback.error.BusinessException;
+import com.programmers.bucketback.error.ErrorCode;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
